backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.135.2
uvicorn==0.42.0
starlette==1.0.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.12.4
pydantic-settings==2.11.0

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.41       # install separately: pip install sqlalchemy
PyMySQL==1.1.1           # install separately: pip install pymysql

# ── Google AI ─────────────────────────────────────────────────────────────────
google-genai==1.68.0
google-generativeai==0.8.6
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-ai-generativelanguage==0.6.15
googleapis-common-protos==1.73.1

# ── HTTP / async ──────────────────────────────────────────────────────────────
anyio==4.13.0
httpx==0.28.1

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv==1.2.1
